+++ CHANGELOG.md
@@ -5,10 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- - `proselint` for `*.{md,txt}` files -->
-<!-- - `npm install --dry-run` for `package.json` files -->
-
 ## \[Unreleased\]
+
+- Added
+  - Add new linters
+    - `proselint` for `*.{md,txt}` files
+    - `npm install --dry-run` for `package.json` files
 
 ## \[0.6.1\] - 2023-06-30